rename fkbeta to fkweb
