+++ packages/fkweb/README.md
@@ -19,7 +19,7 @@ Use pip to install the rest of the requirements:
 
     (env)$ pip install -r requirements-dev.txt
 
-Open fkbeta/settings.py in your editor and edit paths and database settings.
+Open fkweb/settings.py in your editor and edit paths and database settings.
 You should set `SECRET_KEY` to a random string.
 
 Initialize the database:
