+++ packages/fkweb/fkweb/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'fkweb.apps.FrikanalenAppConfig'
