+++ packages/fkweb/fkweb/urls.py
@@ -10,7 +10,7 @@ from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 import agenda.urls
 import api.urls
 import news.urls
-from fkbeta.views import Frontpage
+from fkweb.views import Frontpage
 
 admin.autodiscover()
