+++ packages/fkweb/fkweb/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for fkbeta project.
+WSGI config for fkweb project.
 
 This module contains the WSGI application used by Django's development server
 and any production WSGI deployments. It should expose a module-level variable
@@ -23,7 +23,7 @@ SITE_ROOT = dirname(dirname(abspath(__file__)))
 # if running multiple sites in the same mod_wsgi process. To fix this, use
 # mod_wsgi daemon mode with each site in its own daemon process, or use
 # os.environ["DJANGO_SETTINGS_MODULE"] = "jajaja.settings"
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "fkbeta.settings.production")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "fkweb.settings.production")
 
 def application(environ, start_response):
     # pass the WSGI environment variables on through to os.environ
