+++ packages/fkweb/manage.py
@@ -5,8 +5,8 @@ import os
 import sys
 
 if __name__ == "__main__":
-    local = "fkbeta.settings.local"
-    test = "fkbeta.settings.test"
+    local = "fkweb.settings.local"
+    test = "fkweb.settings.test"
     os.environ.setdefault("DJANGO_SETTINGS_MODULE", test if 'test' in sys.argv else local)
 
     from django.core.management import execute_from_command_line
